refactor: 패키지 구조 변경
* 컨트롤러 클래스 이름 변경
* 각 클래스를 적합한 패키지로 이동
