+++ src/main/java/racingcar/view/Output.java
@@ -1,6 +1,7 @@
-package racingcar.io;
+package racingcar.view;
 
 import racingcar.domain.Car;
+import racingcar.text.Text;
 
 import java.util.List;
