+++ src/test/java/racingcar/view/InputTest.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.view;
 
 import org.junit.jupiter.api.Test;
